AcquisitionManager: move to org.micromanager.acquisition package.
